Inital commit of Issue #168
